Initial version of err
This design deviates from the original in multiple way. Fist, it uses either a battery input or a barrel power input. Second, it uses 24v instead of the original 12v. And, finally, it uses a dedicated LED driver. I am calling this initial design an upgraded v1.0. Will probably keep iterating on it depending on what I see when these boards are built.

Signed-off-by: Nathan Hjelm <hjelmn@me.com>